+++ client/src/components/PermissionsSwitch.js
@@ -3,53 +3,7 @@ import Switch from '@material-ui/core/Switch';
 import List from '@material-ui/core/List';
 import ListItem from '@material-ui/core/ListItem';
 import API from "../utils/API";
-
-const switches = [
-    {
-        label: "Dashboard",
-        bitmask: 0b1,
-    },
-    {
-        label: "Sales Team Daily",
-        bitmask: 0b10,
-    },
-    {
-        label: "Inventory",
-        bitmask: 0b100,
-    },
-    {
-        label: "Client List",
-        bitmask: 0b1000,
-    },
-    {
-        label: "Purchasing Tool",
-        bitmask: 0b10000,
-    },
-    {
-        label: "Sales Analytics",
-        bitmask: 0b100000,
-    },
-    {
-        label: "Manager Task Assignment",
-        bitmask: 0b1000000,
-    },
-    {
-        label: "Sales Team Analytics",
-        bitmask: 0b10000000,
-    },
-    {
-        label: "Check User Messages",
-        bitmask: 0b100000000,
-    },
-    {
-        label: "Discover",
-        bitmask: 0b1000000000,
-    },
-    {
-        label: "Add Or Remove Users",
-        bitmask: 0b10000000000,
-    }
-];
+import switches from '../permissions'
 
 class PermissionsSwitch extends Component {
     state = {
@@ -87,7 +41,7 @@ class PermissionsSwitch extends Component {
                 <List>
                     {switches.map(perm => (
                         <ListItem key={`${perm.bitmask}`}>
-                            {perm.label} <Switch
+                            {perm.title} <Switch
                                 checked={this.isChecked("perms", perm.bitmask)}
                                 onClick={this.handleChange("perms", perm.bitmask)} />
                         </ListItem>
