+++ controllers/messageController.js
@@ -25,6 +25,12 @@ module.exports = {
             .then(dbModel => res.json(dbModel))
             .catch(err => res.status(422).json(err));
     },
+    findUnread: function (req,res) {
+        Message
+            .find({read: false})
+            .then(dbModel => res.json(dbModel))
+            .catch(err => res.status(422).json(err));
+    },
     create: function (req, res) {
         Message
             .create(req.body)
