+++ package.json
@@ -20,7 +20,6 @@
   "license": "ISC",
   "devDependencies": {
     "concurrently": "^4.1.2",
-    "dotenv": "^8.1.0",
     "eslint": "^6.3.0",
     "nodemon": "^1.19.1"
   },
@@ -31,6 +30,7 @@
     "bcryptjs": "^2.4.3",
     "body-parser": "^1.19.0",
     "date-fns": "^2.0.0-beta.5",
+    "dotenv": "^8.1.0",
     "express": "^4.17.1",
     "google-maps-react": "^2.0.2",
     "http-proxy-middleware": "^0.19.1",
